+++ src/Connectivity/Shared/CrossConnectivity.shared.cs
@@ -33,7 +33,7 @@ namespace Plugin.Connectivity
 
         static IConnectivity CreateConnectivity()
         {
-#if NETSTANDARD1_0
+#if NETSTANDARD1_0 || NETSTANDARD2_0
             return null;
 #else
 			return new ConnectivityImplementation();